third_party/WebKit/LayoutTests/battery-status/multiple-promises-after-resolve-expected.txt

Test battery status API with multiple promises after resolve.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


resolution number 1
PASS batteryInfo is defined.
PASS mockBatteryInfo is defined.
PASS batteryInfo.charging is mockBatteryInfo.charging
PASS batteryInfo.chargingTime is mockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is mockBatteryInfo.dischargingTime
PASS batteryInfo.level is mockBatteryInfo.level
PASS promise1 === promise2 is true
resolution number 2
PASS batteryInfo is defined.
PASS mockBatteryInfo is defined.
PASS batteryInfo.charging is mockBatteryInfo.charging
PASS batteryInfo.chargingTime is mockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is mockBatteryInfo.dischargingTime
PASS batteryInfo.level is mockBatteryInfo.level
PASS successfullyParsed is true

TEST COMPLETE
